requirements.txt
dnslib>=0.9.24
pyyaml>=6.0.1
